app/requirements.txt
fastapi
uvicorn[standard]
pandas
supabase
openai
python-dotenv
python-multipart
numpy
pandasai
